backend/companion_app/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
crawl4ai
asyncio
python-multipart
